Leaf mass per area and photosynthetic rate for plant species from different sites

This dataset contains the leaf mass per area (LMA) and photosynthetic rate per meaf mass (Amass) of individual plants from 81 plant species at four sites in south-eastern Australia. Sites represent contrasts along rainfall and soil phosphorous gradients.

leafmeas

Format

A data frame containing 529 rows and 4 variables:

site

(factor) The site at which this species was sampled (coded as Murrua, WH, RHM, RHW)

spp

(factor) Species of the observed plant species

Amass

(numeric) Photosynthetic rate per leaf mass ()

lma

(numeric) Leaf mass per area (m2/kg)

Source

Wright I. J., Reich P. B. and Westoby M. (2001) Strategy shifts in leaf physiology, structure and nutrient content between species of high- and low-rainfall and high- and low-nutrient habitats. Functional Ecology 15, 423--434.